Introduce flat-top windowing and add unit tests
Improve the CNR estimation by introducing windowing, particularly using a
flat-top window. The windowing avoids the strong leakage that was present in the
previous implementation (non-windowed, i.e., based on a rectangular/boxcar
window). Furthermore, the flat-top window avoids problems with scalloping losses
when the beacon frequency is not aligned with an FFT bin.

Update the beacon sink block API to support testing. Add methods to fetch the
last CNR and CW beacon frequency estimates. Use those methods to verify the CNR
estimation in unit tests. To support the new "get_freq()" method, add the sample
rate as a parameter of the beacon sink block, so that the block can return the
correct beacon frequency in Hz.

Remove the previous tracking regarding the average amplitude. It was not useful
given that the amplitude translation to dBm is unknown (it depends on the Rx
hardware).

Based on the experiments in unit tests regarding the new windowed
implementation, change the default alpha coefficient used for
exponentially-weighted moving average (EWMA). Set it to 0.0005 by default (i.e.,
average over 2000 FFT blocks). Within the unit tests, this coefficient
guarantees an estimate within 0.2 dB of the real CNR.

Lastly, add some matlab sources used to analyze the windowing process and
particularly to find the appropriate window function for the beacon CNR
estimation use case.