Fix frequency estimate by wrapping FFT peak index
The previous estimate was based on the peak average FFT squared
magnitude before an fftshift/wrapping operation. As a result, negative
frequencies (below the center frequency) would show up on index > N/2
for an N-point FFT and the printed beacon frequency should be > fs/2.
Fix this by wrapping the peak index properly.